requirements.txt
APScheduler==3.6.0
Click==7.0
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-APScheduler==1.11.0
Flask-SQLAlchemy==2.3.2
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
redis==3.2.1
six==1.12.0
SQLAlchemy==1.3.1
tzlocal==1.5.1
Werkzeug==0.15.1
Pillow==6.0.0
PyJWT==1.7.1
XlsxWriter==1.1.8
docxtpl==0.5.17
aliyun-python-sdk-core==2.13.4